backend/list_gemini_models.py
import google.generativeai as genai
import os

API_KEY = os.getenv("GEMINI_API_KEY")
if not API_KEY:
    raise RuntimeError("GEMINI_API_KEY not set")

genai.configure(api_key=API_KEY)

print("Available Gemini models:")
for m in genai.list_models():
    print(m.name)
